refactor(pull): split hierarchical and adhoc pull into separate use cases
Separate goalStateByWeek hierarchy pulls from adhoc week-indexed pulls so each path has its own logic, probes, and tests while preserving the combined API contract.
