--- services/backend/src/usecase/moveGoalsFromWeek/index.ts
@@ -1,1 +1,0 @@
-export * from './moveGoalsFromWeek';
